# Conflicts:
#	commander/apply_transactions.go
#	commander/apply_transactions_test.go

+++ db/migrations/000003_transaction.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "transaction" (
     tx_hash                BYTEA PRIMARY KEY,
-    from_index             NUMERIC(78) NOT NULL,
-    to_index               NUMERIC(78) NOT NULL,
+    from_index             BIGINT NOT NULL,
+    to_index               BIGINT NOT NULL,
     amount                 NUMERIC(78) NOT NULL,
     fee                    NUMERIC(78) NOT NULL,
     nonce                  NUMERIC(78) NOT NULL,
